change data type of buy_price in migration

+++ server/db/migrations/20210625205425_portfolio.js
@@ -3,7 +3,7 @@ exports.up = function (knex) {
     table.increments('id').primary()
     table.string('name')
     table.string('ticker')
-    table.integer('buy_price')
+    table.decimal('buy_price')
     table.integer('number_shares')
   })
 }
